+++ zcash_client_backend/CHANGELOG.md
@@ -15,6 +15,10 @@ workspace.
 ### Added
 - `zcash_client_backend::tor::http::cryptex`:
     - `exchanges::{CoinEx, DigiFinex, Kraken, Xt}`
+- `zcash_client_backend::wallet::Exposure`
+- `zcash_client_backend::wallet::TransparentAddressSource`
+- `zcash_client_backend::wallet::TransparentAddressMetadata::derived`
+- `zcash_client_backend::wallet::TransparentAddressMetadata::exposure`
 
 ### Changed
 - Migrated to `zcash_protocol 0.7`, `zcash_address 0.10`, `zip321 0.6`,
@@ -22,6 +26,14 @@ workspace.
 - `zcash_client_backend::data_api`:
   - `testing::pool::ShieldedPoolTester` has added methods `note_value` and
     `select_unspent_notes`.
+- `zcash_client_backend::wallet`:
+  - `TransparentAddressMetadata` has been converted from an enum to a struct
+    that contains both source metadata and information about when the address
+    was exposed by the wallet. The derivation information that this type
+    previously provided is now provided by `TranparentAddressSource`. As a
+    consequence of this change, the signature of
+    `TransparentAddressMetadata::new` has changed; use
+    `TransparentAddressMetadata::derived` instead.
 
 ### Removed
 - `zcash_client_backend::tor::http::cryptex::exchanges::GateIo`